Update upload controller unit tests
